+++ client/src/components/common/ClickableContainer.tsx
